Move new boxes to library.
